+++ src/rio.h
@@ -167,7 +167,7 @@ void rioInitWithFd(rio *r, int fd);
 void rioFreeFd(rio *r);
 void rioFreeConn(rio *r, sds* out_remainingBufferedData);
 
-size_t rioWriteBulkCount(rio *r, char prefix, long count);
+size_t rioWriteBulkCount(rio *r, char prefix, long long count);
 size_t rioWriteBulkString(rio *r, const char *buf, size_t len);
 size_t rioWriteBulkLongLong(rio *r, PORT_LONGLONG l); 			WIN_PORT_FIX
 size_t rioWriteBulkDouble(rio *r, double d);
